Move modflow methods and tests
